+++ app.rb
@@ -15,10 +15,7 @@ require 'sinatra/flash'
 #require 'rack/csrf' # TODO: resolver esto, es importante tener un csrf por motivos de seguridad en los formularios.
 #require 'sinatra/handlebars'
 
-require_relative 'libs/redmine_user.rb'
-require_relative 'libs/redmine_issues.rb'
-require_relative 'libs/redmine_journals.rb'
-require_relative 'libs/redmine_connect2api.rb'
+require_relative 'libs/redmine.rb'
 require_relative 'libs/issues.rb'
 require_relative 'helpers/init'
