+++ libs/redmine.rb
@@ -1,0 +1,4 @@
+Dir[File.dirname(__FILE__) + '/redmine/*.rb'].each do |file|
+    require file
+end
+
