+++ libs/redmine/project.rb
@@ -1,0 +1,11 @@
+module Redmine
+  class Project
+    def initialize()
+      @config = YAML.load_file("./conf/config.yaml")
+    end
+
+    def get_issues_by_project_id() 
+
+    end
+  end
+end
