--- libs/redmine_project.rb
@@ -1,9 +1,0 @@
-def RedmineProject
-  def initialize()
-    @config = YAML.load_file("./conf/config.yaml")
-  end
-
-  def get_issues_by_project_id() 
-
-  end
-end
