+++ routes/user.rb
@@ -8,7 +8,7 @@ class SmartRedmine < Sinatra::Base
       redirect '/projects/' + cookies[:project_id]
     end
 
-    response = RedmineUser.new.get_projects(session[:user]['api_key'])
+    response = Redmine::User.new.get_projects(session[:user]['api_key'])
 
     if !response.nil?
       @projects = response
@@ -17,7 +17,6 @@ class SmartRedmine < Sinatra::Base
     end
 
     erb  :'../views/user'
-
   end
-
 end
+
